+++ src/components/index.ts
@@ -6,5 +6,6 @@ import Title from './title'
 import TextField from './form/text-field'
 import Image from './image'
 import IconButton from './icon-button'
+import Dialog from './dialog'
 
-export { Button, Icon, IconButton, PasswordField, Text, TextField, Title, Image }
+export { Button, Icon, IconButton, PasswordField, Text, TextField, Title, Image, Dialog }